feat: Add Vibe Coder chat input and message components with updated styles and functionality
