+++ experiments/vibe-coder/vibe-coder-header.js
@@ -3,6 +3,7 @@ class VibeCoderHeader extends HTMLElement {
         super();
         this.attachShadow({ mode: 'open' });
         this.shadowRoot.innerHTML = `
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/7.0.1/css/all.min.css">
             <style>
                 header {
                     height: 3.5rem;
